backend/requirements.txt
asgiref==3.8.1
autopep8==2.3.2
diff-match-patch==20241021
dj-database-url==3.0.0
Django==4.2.21
django-cors-headers==4.7.0
django-import-export==4.3.8
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
gunicorn==23.0.0
more-itertools==10.7.0
numpy==2.0.2
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
psycopg2-binary==2.9.10
pycodestyle==2.13.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.14.0
tzdata==2025.2
whitenoise==6.9.0
